#285 Rename Neg to Not
